requirements.txt
streamlit>=1.38.0
google-genai>=1.0.0
openpyxl>=3.1.0
python-docx>=1.1.0
PyPDF2>=3.0.0
pandas>=2.0.0
